Create post-play.cna
This script can be usefull for fast and very basic post-exploitation automation task. 
By using it, you are able to define a command list that will be execute once 
a meterpreter or shell session are created. You also have the possibility to run quickly a list
of command again a specific target. Commands are stored in a text file (one per line).

it work like resource file or autorunscript's option of meterpreter but with more malleability.