Major update of BeefStrike
There is to many thing to say about this update. i will put a video that
will speak at my place.
But i can say that this one is more stable are open new vista like:
Automatic and targeted browser's exploitation script. Based on their
profiles. a PoC script is in progress too